Seperate CentralWidget class to different file
